test: init integration tests

+++ src/main/java/br/com/itau/transferapi/application/req/TransactionParams.java
@@ -3,8 +3,6 @@ package br.com.itau.transferapi.application.req;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Getter
 @RequiredArgsConstructor
 public class TransactionParams {
@@ -12,5 +10,5 @@ public class TransactionParams {
   private final String originWalletId;
   private final String targetClientId;
   private final String targetWalletId;
-  private final BigDecimal amount;
+  private final double amount;
 }
